updated navbar, splitted in components
