+++ components/Layout.js
@@ -1,13 +1,11 @@
 import Footer from 'components/Footer'
-import Header from 'components/Header'
-import Navbar from 'components/Navbar'
+import Navbar from 'components/Navbar/Navbar'
 import PropTypes from 'prop-types'
 import styles from '../styles/Home.module.css'
 
 const Layout = ({ children }) => {
   return (
     <div className={styles.container}>
-      {/* <Header /> */}
       <Navbar />
       <main className={styles.main}>{children}</main>
       <Footer />
